+++ packages/antd-management-fast-component/src/components/Function/Assist/index.js
@@ -1,3 +1,5 @@
+import { ColorFactory } from 'antd/es/color-picker/color';
+
 /**
  * convert option or radio data
  * @param {*} o data source
@@ -13,6 +15,30 @@ export function convertOptionOrRadioData(o, index) {
     ...o,
   };
 
+  return {
+    index,
+    label: name,
+    value: flag,
+    disabled: !availability,
+    ...o,
+  };
+}
+
+/**
+ * convert option or radio data
+ * @param {*} o data source
+ * @param {*} index data index
+ * @returns
+ */
+export function convertOptionOrRadioData1(o, index) {
+  const { flag, name, availability } = {
+    flag: `option_${index}`,
+    name: '',
+    availability: 1,
+    hidden: false,
+    ...o,
+  };
+  ColorFactory;
   return {
     index,
     label: name,