+++ packages/antd-management-fast-component/src/components/FunctionComponent/ComponentBuilder/index.jsx
@@ -1,4 +1,4 @@
-import { Alert, Avatar, Button, Dropdown, List, Tree } from 'antd';
+import { Alert, Avatar, Button, ColorPicker, Dropdown, List, Tree } from 'antd';
 import React from 'react';
 import ReactPlayer from 'react-player';
 
@@ -21,6 +21,7 @@ import { ElasticityMenuHeader } from '../../ElasticityMenuHeader';
 import { ElasticityRadioGroup } from '../../ElasticityRadioGroup';
 import { ElasticityTagList } from '../../ElasticityTagList';
 import { ElasticityTreeSelect } from '../../ElasticityTreeSelect';
+import { FlexBox } from '../../FlexBox';
 import { FlexRadio } from '../../FlexRadio';
 import { FlexSelect } from '../../FlexSelect';
 import { FlexText } from '../../FlexText';
@@ -28,6 +29,7 @@ import { iconBuilder } from '../../Icon';
 import { IconInfo } from '../../IconInfo';
 import { JsonView } from '../../JsonView';
 import { StatusBar } from '../../StatusBar';
+import { VerticalBox } from '../../VerticalBox';
 
 const ButtonGroup = Button.Group;
 
@@ -400,6 +402,72 @@ export function buildRadioGroup({
       list={list}
       adjustListDataCallback={adjustListDataCallback}
       onChange={onChange}
+    />
+  );
+}
+
+export function buildColorPicker({
+  value = null,
+  disabled = false,
+  onChange = null,
+}) {
+  return (
+    <FlexBox
+      flexAuto="right"
+      left={<VerticalBox>色值：</VerticalBox>}
+      right={
+        <div
+          style={{
+            minWidth: '104px',
+          }}
+        >
+          <VerticalBox>
+            <ColorPicker
+              value={value}
+              showText
+              disabled={disabled}
+              presets={[
+                {
+                  label: '常用',
+                  colors: [
+                    '#000000',
+                    '#000000E0',
+                    '#000000A6',
+                    '#00000073',
+                    '#00000040',
+                    '#00000026',
+                    '#0000001A',
+                    '#00000012',
+                    '#0000000A',
+                    '#00000005',
+                    '#F5222D',
+                    '#FA8C16',
+                    '#FADB14',
+                    '#8BBB11',
+                    '#52C41A',
+                    '#13A8A8',
+                    '#1677FF',
+                    '#2F54EB',
+                    '#722ED1',
+                    '#EB2F96',
+                    '#F5222D4D',
+                    '#FA8C164D',
+                    '#FADB144D',
+                    '#8BBB114D',
+                    '#52C41A4D',
+                    '#13A8A84D',
+                    '#1677FF4D',
+                    '#2F54EB4D',
+                    '#722ED14D',
+                    '#EB2F964D',
+                  ],
+                },
+              ]}
+              onChange={onChange}
+            />
+          </VerticalBox>
+        </div>
+      }
     />
   );
 }