+++ packages/antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 2.11.159
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.6.142
+  - antd-management-fast-common@2.6.93
+
 ## 2.11.158
 
 ### Patch Changes