+++ packages/example-antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 1.8.171
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.6.142
+  - antd-management-fast-framework@2.11.159
+  - antd-management-fast-common@2.6.93
+
 ## 1.8.170
 
 ### Patch Changes